backend_new/requirements.txt
# Backend dependencies for the Survey Bot application
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
qrcode==7.4.2
pandas==2.1.4
openpyxl==3.1.2
reportlab==4.0.5
pillow==10.3.0
PyMySQL==1.1.0
cryptography==42.0.8